requirements.txt
streamlit==1.27.0
langchain==0.0.310
python-dotenv==1.0.0
pypdf==3.16.2
rapidocr-onnxruntime==1.3.7
openai==0.28.1
moviepy==1.0.3
